Animations & hitboxes for ice skater
-Player also transitions into their crouching animation

+++ Assets/_Project/_FighterBase/_Scripts/Hitboxes/Hitbox.cs
@@ -27,9 +27,9 @@ public class Hitbox : MonoBehaviour
     private HashSet<Hurtbox> alreadyHit = new HashSet<Hurtbox>();
     private Hurtbox hurtbox;
     [SerializeField] private Collider2D ownerHurtboxCollider;
-    private float hitstunFrames;
+    private int hitstunFrames;
 
-    public float HitstunFrames { get { return hitstunFrames; }  set { hitstunFrames = value; } }
+    public int HitstunFrames { get { return hitstunFrames; }  set { hitstunFrames = value; } }
    
     void Awake()
     {
